Success responses from both SagePay and DataCash simulator APIs.
I've turned the test code 'inside out' so the actual tests are defined in a base PurchaseGatewayTests class that uses an abstract method to create the payment gateway; derived classes for each provider then implement the CreateGateway method.

This seemed like a grand idea at the time but has the horrible drawback that all your actual tests end up in an abstract class so you can no longer run them in the debugger.

I'll probably rethink this tomorrow. But for now, there are tests and they are green.