Added Basket and LineItem classes, because the SagePay gateway requires these to be present
For now, it automatically generates a single line item with a description and total only, where the description is set to "Transaction ref {merchantReference}" - that doesn't actually itemise the order (yet) but it provides enough data for SagePay to process the transaction.

The Basket and LineItem classes should provide a flexible foundation for future improvements to the basket/item/order contents system.